feat: build binary in docker when building image
This makes us always rebuild the binary when building the image, all in
docker.  This ensures that we have better control over the inputs for that
build.

The base image is still require for running, as (at least for now) the
resulting binary depends on the dynamic linker.

Also includes support to vendor any dependencies if we have any `replace`
lines in `go.mod`; this is useful if we're using a locally edited version
of those dependencies (the most likely of which is eirinix).