fix: Build image based on scratch
If we turn off CGO for the binary (which doesn't appear to be needed),
we can get a static binary that then doesn't need the dynamic linker
(and therefore can bse based on an empty image).  Note that we _do_
expect a working /tmp, so we will need to create that in the iamge
first.